+++ cmake/Findossim-registration-source.cmake
@@ -3,6 +3,21 @@ if(TARGET ossim_registration_source::ossim-registration-source)
   return()
 endif()
 
+if(TARGET ossim_autoreg::ossim-registration-source)
+  add_library(ossim_registration_source::ossim-registration-source INTERFACE IMPORTED)
+  set_target_properties(ossim_registration_source::ossim-registration-source PROPERTIES
+    INTERFACE_LINK_LIBRARIES ossim_autoreg::ossim-registration-source)
+  set(ossim-registration-source_FOUND TRUE)
+  return()
+endif()
+
+if(TARGET ossim-registration-source)
+  add_library(ossim_registration_source::ossim-registration-source ALIAS
+    ossim-registration-source)
+  set(ossim-registration-source_FOUND TRUE)
+  return()
+endif()
+
 set(_OSSIM_REGISTRATION_SOURCE_HINTS)
 if(DEFINED OSSIM_REGISTRATION_SOURCE_ROOT)
   list(APPEND _OSSIM_REGISTRATION_SOURCE_HINTS "${OSSIM_REGISTRATION_SOURCE_ROOT}")
@@ -10,7 +25,14 @@ endif()
 if(DEFINED ENV{OSSIM_REGISTRATION_SOURCE_ROOT})
   list(APPEND _OSSIM_REGISTRATION_SOURCE_HINTS "$ENV{OSSIM_REGISTRATION_SOURCE_ROOT}")
 endif()
+if(DEFINED OSSIM_AUTOREG_ROOT)
+  list(APPEND _OSSIM_REGISTRATION_SOURCE_HINTS "${OSSIM_AUTOREG_ROOT}")
+endif()
+if(DEFINED ENV{OSSIM_AUTOREG_ROOT})
+  list(APPEND _OSSIM_REGISTRATION_SOURCE_HINTS "$ENV{OSSIM_AUTOREG_ROOT}")
+endif()
 list(APPEND _OSSIM_REGISTRATION_SOURCE_HINTS
+  "${CMAKE_CURRENT_LIST_DIR}/../../ossim-autoreg"
   "${CMAKE_CURRENT_LIST_DIR}/../../ossim-registration-source")
 
 find_path(OSSIM_REGISTRATION_SOURCE_INCLUDE_DIR